GH-27 Change files naming and controller responsibility
