+++ src/components/EmptyState/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EmptyState';
