+++ src/pages/Applications/List/ApplicationsList.queries.ts
