+++ src/pages/Applications/List/ApplicationsList.styles.css.ts
