--- src/pages/Applications/List/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ApplicationsListController';
