+++ src/pages/Generator/Generator.tsx
